feat: :sparkles: option to get product

+++ elevate/api/views.py
@@ -3,6 +3,7 @@ from django.shortcuts import render
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework.status import HTTP_400_BAD_REQUEST
+
 from django.shortcuts import get_object_or_404
 from .models import Product
 from .serializers import productSerializer
@@ -26,9 +27,14 @@ def product_list(request):
         else:
             return Response(serializer.errors,status=HTTP_400_BAD_REQUEST)
 
-    
-@api_view(['DELETE'])
-def delete_product(request,productId):
-    count=get_object_or_404(Product,id=productId).delete()
-    return Response({'deleted':count[0]})
+@api_view(['GET','PUT','DELETE'])
+def product(request,productId):
+    if request.method == 'GET':
+        product=get_object_or_404(Product,id=productId)
+        serializer=productSerializer(product)
+        return Response(serializer.data) 
+    if request.method == 'DELETE':
+        count=get_object_or_404(Product,id=productId).delete()
+        return Response({'deleted':count[0]})
 
+
